added; tests for bluebird and q
