+++ lib/index.js
@@ -602,7 +602,6 @@ Mongoose.prototype.Query = Query;
  */
 
 Object.defineProperty(Mongoose.prototype, 'Promise', {
-  writable: true,
   get: function() {
     return PromiseProvider.get();
   },
