+++ test/promise.test.js
@@ -4,11 +4,8 @@
  */
 
 var assert = require('assert');
-var start = require('./common');
 
 var Promise = require('../lib/promise');
-var PromiseProvider = require('../lib/promise_provider');
-var Schema = require('../lib/schema');
 
 /**
  * Test.
@@ -210,162 +207,6 @@ describe('Promise', function(){
         });
 
         p.reject(3);
-      });
-    });
-  });
-});
-
-describe('ES6 promises: ', function() {
-  describe('native: ', function() {
-    if (!global.Promise) {
-      return;
-    }
-
-    var Promise;
-    var db;
-    var testSchema = new Schema({ test: { type: String, required: true } });
-    testSchema.pre('save', function(next) {
-      if (this.$__saveSucceeds === false) {
-        return next(new Error('fail'));
-      }
-      next();
-    });
-    testSchema.pre('validate', function(next) {
-      if (this.$__validateSucceeds === false) {
-        return next(new Error('validation failed'));
-      }
-      next();
-    });
-    testSchema.pre('findOne', function(next) {
-      if (this.$__findOneSucceeds === false) {
-        return next(new Error('findOne failed'));
-      }
-      next();
-    });
-    var MyModel;
-
-    before(function() {
-      PromiseProvider.set(global.Promise);
-      Promise = PromiseProvider.get();
-    });
-
-    before(function() {
-      db = start();
-      MyModel = db.model('es6promise', testSchema);
-    });
-
-    after(function() {
-      PromiseProvider.reset();
-    });
-
-    afterEach(function(done) {
-      MyModel.remove({}, done);
-    });
-
-    it('save()', function(done) {
-      var m = new MyModel({ test: '123' });
-      var promise = m.save();
-      assert.equal(promise.constructor, global.Promise);
-      promise.then(function(doc) {
-        assert.equal(m, doc);
-        done();
-      });
-    });
-
-    it('save() with validation error', function(done) {
-      var m = new MyModel({});
-      var promise = m.save();
-      assert.equal(promise.constructor, global.Promise);
-      promise.
-        then(function(doc) {
-          assert.ok(false);
-        }).
-        catch(function(err) {
-          assert.ok(err);
-          assert.ok(err.errors['test']);
-          done();
-        });
-    });
-
-    it('save() with middleware error', function(done) {
-      var m = new MyModel({ test: '123' });
-      m.$__saveSucceeds = false;
-      var promise = m.save();
-      assert.equal(promise.constructor, global.Promise);
-      promise.
-        then(function(doc) {
-          assert.ok(false);
-        }).
-        catch(function(err) {
-          assert.ok(err);
-          assert.equal(err.toString(), 'Error: fail');
-          done();
-        });
-    });
-
-    it('save() with validation middleware error', function(done) {
-      var m = new MyModel({ test: '123' });
-      m.$__validateSucceeds = false;
-      var promise = m.save();
-      assert.equal(promise.constructor, global.Promise);
-      promise.
-        then(function(doc) {
-          assert.ok(false);
-        }).
-        catch(function(err) {
-          assert.ok(err);
-          assert.equal(err.toString(), 'Error: validation failed');
-          done();
-        });
-    });
-
-    it('validate()', function(done) {
-      var m = new MyModel({});
-      var promise = m.validate();
-      assert.equal(promise.constructor, global.Promise);
-      promise.
-        then(function(doc) {
-          assert.ok(false);
-        }).
-        catch(function(err) {
-          assert.ok(err);
-          assert.ok(err.errors['test']);
-          done();
-        });
-    });
-
-    it('queries', function(done) {
-      MyModel.create({ test: '123' }, function(error) {
-        assert.ifError(error);
-
-        var promise = MyModel.findOne({ test: '123' }).exec();
-        assert.equal(promise.constructor, global.Promise);
-
-        promise.then(function(doc) {
-          assert.equal(doc.test, '123');
-          done();
-        });
-      });
-    });
-
-    it('queries with errors', function(done) {
-      MyModel.create({ test: '123' }, function(error) {
-        assert.ifError(error);
-
-        var query = MyModel.findOne({ test: '123' });
-        query.$__findOneSucceeds = false;
-        var promise = query.exec();
-        assert.equal(promise.constructor, global.Promise);
-
-        promise.
-          then(function() {
-            assert.ok(false);
-          }).
-          catch(function(err) {
-            assert.ok(err);
-            assert.equal(err.toString(), 'Error: findOne failed');
-            done();
-          });
       });
     });
   });
